+++ src/extension.ts
@@ -15,7 +15,7 @@ import {
   markdownToHtml,
   markdownToPptx,
   markdownToTextile,
-} from "./markdown-docx/markdown-to-docx";
+} from "./markdown-docx/markdown-to-xxxx";
 // import { markdownToPptx } from "./markdown-pptx/markdown-to-pptx";
 import { createDocxTemplateFile } from "./markdown-docx/common";
 import { wordDownToPptx } from "./markdown-docx/wd-to-pptx";
